Move tesla_wall_connector base entity to separate module
